batch-worker/requirements.txt
Django>=3.1,<4.0
celery~=5.5.3
redis
psycopg2-binary
django-celery-results
django-celery-beat
dj-database-url
flower
python-logging-loki==0.3.1
python-dotenv==1.0.1
tzlocal~=5.2
pytz~=2024.1
requests~=2.32.3
python-dateutil~=2.9.0.post0
pyyaml
python-loki-logger
concurrent-log-handler
greenlet==3.0.3
gevent==24.2.1

opentelemetry-api>=1.35.0
opentelemetry-sdk>=1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-instrumentation-django>=0.56b0
opentelemetry-instrumentation-celery>=0.56b0
opentelemetry-exporter-otlp>=1.35.0
opentelemetry-instrumentation-logging>=0.56b0
opentelemetry-instrumentation-psycopg2>=0.56b0
opentelemetry-instrumentation-redis>=0.56b0
opentelemetry-instrumentation-requests>=0.56b0
opentelemetry-instrumentation-sqlite3>=0.56b0
opentelemetry-instrumentation-urllib>=0.56b0
opentelemetry-instrumentation-urllib3>=0.56b0